homework_06/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
psycopg = "^3.2.2"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-wtf = "^1.2.1"
wtforms = {extras = ["email"], version = "^3.1.2"}

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.3"
